refactor(cli): deduplicate shared utilities into common modules
Extract duplicated functions into shared modules:

- formatJsoncParseError -> utils/jsonc.ts
- parseVersion, compareSemver -> utils/semver.ts
- hashContent consolidated in utils/receipt.ts
- validation error factories -> lib/validation-errors-factory.ts
- inline isPlainObject replaced with utils/type-guards import
