+++ packages/cli/src/commands/build.ts
@@ -8,21 +8,17 @@ import { relative, resolve } from "node:path"
 import type { Command } from "commander"
 import kleur from "kleur"
 import { BuildRegistryError, type BuildRegistryResult, buildRegistry } from "../lib/build-registry"
-import { runCompleteValidation } from "../lib/validation-runner"
-import type { LoadRegistryErrorKind } from "../lib/validators"
-import { outputDryRun } from "../utils/dry-run"
 import {
-	EXIT_CODES,
-	NotFoundError,
-	OCXError,
-	ValidationFailedError,
-	type ValidationFailureDetails,
-} from "../utils/errors"
+	createLoadValidationError,
+	createValidationFailureError,
+} from "../lib/validation-errors-factory"
+import { runCompleteValidation } from "../lib/validation-runner"
+import { outputDryRun } from "../utils/dry-run"
 import { handleError } from "../utils/handle-error"
 import { outputJson } from "../utils/json-output"
 import { logger } from "../utils/logger"
 import { createSpinner } from "../utils/spinner"
-import { categorizeValidationErrors, summarizeValidationErrors } from "../utils/validation-errors"
+import { categorizeValidationErrors } from "../utils/validation-errors"
 
 interface BuildOptions {
 	cwd: string
@@ -31,43 +27,6 @@ interface BuildOptions {
 	quiet: boolean
 	dryRun?: boolean
 	showValidation: boolean
-}
-
-function createLoadValidationError(message: string, errorKind?: LoadRegistryErrorKind): Error {
-	if (errorKind === "not_found") {
-		return new NotFoundError(message)
-	}
-
-	if (errorKind === "parse_error") {
-		return new OCXError(message, "CONFIG_ERROR", EXIT_CODES.CONFIG)
-	}
-
-	return new OCXError(message, "CONFIG_ERROR", EXIT_CODES.CONFIG)
-}
-
-function createValidationFailureError(
-	errors: string[],
-	failureType: "schema" | "rules",
-): ValidationFailedError {
-	const summary = summarizeValidationErrors(errors, {
-		schemaErrors: failureType === "schema" ? errors.length : 0,
-	})
-
-	const details: ValidationFailureDetails = {
-		valid: false,
-		errors,
-		summary: {
-			valid: false,
-			totalErrors: summary.totalErrors,
-			schemaErrors: summary.schemaErrors,
-			sourceFileErrors: summary.sourceFileErrors,
-			circularDependencyErrors: summary.circularDependencyErrors,
-			duplicateTargetErrors: summary.duplicateTargetErrors,
-			otherErrors: summary.otherErrors,
-		},
-	}
-
-	return new ValidationFailedError(details)
 }
 
 function outputValidationFailures(errors: string[], failureType: "schema" | "rules"): void {
