+++ packages/cli/src/profile/manager.ts
@@ -1,5 +1,5 @@
 import { mkdir, readdir, rename, rm, stat } from "node:fs/promises"
-import { type ParseError, parse as parseJsonc, printParseErrorCode } from "jsonc-parser"
+import { type ParseError, parse as parseJsonc } from "jsonc-parser"
 import type { ProfileOcxConfig } from "../schemas/ocx"
 import { profileOcxConfigSchema } from "../schemas/ocx"
 import {
@@ -11,6 +11,7 @@ import {
 	ProfileNotFoundError,
 	ProfilesNotInitializedError,
 } from "../utils/errors"
+import { formatJsoncParseError } from "../utils/jsonc"
 import { atomicWrite } from "./atomic"
 import {
 	getLocalProfileDir,
@@ -60,19 +61,6 @@ export const DEFAULT_OCX_CONFIG_TEMPLATE = `{
   "include": []
 }
 `
-
-function formatJsoncParseError(parseErrors: ParseError[]): string {
-	if (parseErrors.length === 0) {
-		return "Unknown parse error"
-	}
-
-	const firstError = parseErrors[0]
-	if (!firstError) {
-		return "Unknown parse error"
-	}
-
-	return `${printParseErrorCode(firstError.error)} at offset ${firstError.offset}`
-}
 
 function parseJsoncOrThrow(content: string, filePath: string): unknown {
 	const parseErrors: ParseError[] = []
