+++ packages/cli/src/utils/index.ts
@@ -9,8 +9,10 @@ export * from "./git-context"
 export * from "./handle-error"
 export * from "./instruction-paths"
 export * from "./json-output"
+export * from "./jsonc"
 export * from "./logger"
 export * from "./path-helpers"
+export * from "./semver"
 export * from "./shared-options"
 export * from "./spinner"
 export * from "./url"
